Struct holmes::engine::Engine

pub struct Engine {
    // some fields omitted
}

The Engine type contains the context necessary to run a Holmes program

Methods

impl Engine

fn new(db: Box<FactDB>) -> Engine

Create a fresh engine by handing it a fact database to use

fn get_type(&self, name: &str) -> Option<Type>

Seach the type registry for a named type If present, it returns Some(type), otherwise None

fn add_type(&mut self, type_: Type) -> Result<(), Error>

Register a new type This type must be a named type (e.g. type.name() should return Some)

fn new_predicate(&mut self, pred: &Predicate) -> Result<(), Error>

Register a new predicate This defines the type signature of a predicate and persists it

Predicates must have at least one argument

Predicates must have a unique name

While using the pg backend, their name must be lowercase ascii or '_'

fn new_fact(&mut self, fact: &Fact) -> Result<(), Error>

Adds a new fact to the database If the fact is already present, a new copy will not be added.

The relevant predicate must already be registered

The fact must be correctly typed

fn derive(&self, query: &Vec<Clause>) -> Result<Vec<Vec<Value>>, Error>

Given a query (similar to the rhs of a rule in Datalog), provide the set of satisfying answers in the database.

fn new_rule(&mut self, rule: &Rule) -> Result<(), Error>

Register a new rule with the database

fn reg_func(&mut self, name: String, func: Func)

Register a new function with the database, to be called from within a rule

Do not attempt to register a function name multiple times.