Struct holmes::engine::types::Field

pub struct Field {
    pub name: Option<String>,
    pub description: Option<String>,
    pub type_: Type,
}

Field for use in a predicate The name is for use in selective matching or unordered definition, and the description is to improve readability of code and comprehension of results. The Type is the only required component of a field, as it defines how to actually interact with the field.

Fields
name: Option<String>

Name of field, for use in matching and instantiating predicates

description: Option<String>

Description of the field, purely documentation, not mechanical

type_: Type

Type of the predicate, explaining how to store and retrieve information from the FactDB