Struct holmes::engine::types::Predicate

pub struct Predicate {
    pub name: String,
    pub description: Option<String>,
    pub fields: Vec<Field>,
}

A Predicate is a name combined with a list of typed slots, e.g.

foo(uint64, string)

would be represented as

use holmes::pg::dyn::types;
use holmes::engine::types::{Predicate, Field};
use std::sync::Arc;
Predicate {
    name: "foo".to_string(),
    description: None,
    fields: vec![Field {
        name: None,
        description: None,
        type_: Arc::new(types::UInt64)
    }, Field {
        name: None,
        description: None,
        type_: Arc::new(types::String)
    }]
};

Fields
name: String

Predicate Name

description: Option<String>

Description of what it means for this predicate to be true. Purely documentation, not mechanical

fields: Vec<Field>

Predicate fields