Trait holmes::fact_db::FactDB

pub trait FactDB {
    fn insert_fact(&mut self, fact: &Fact) -> Result<bool>;
    fn add_type(&mut self, type_: Type) -> Result<()>;
    fn get_type(&self, type_str: &str) -> Option<Type>;
    fn get_predicate(&self, pred_name: &str) -> Option<&Predicate>;
    fn new_predicate(&mut self, pred: &Predicate) -> Result<()>;
    fn search_facts(&self, query: &Vec<Clause>) -> Result<Vec<Vec<Value>>>;
}

This is the interface which a fact database must satisfy to be used by the Holmes engine.

Required Methods

fn insert_fact(&mut self, fact: &Fact) -> Result<bool>

Adds a new fact to the database, returning false if the fact was already present in the database, and true if it was inserted.

fn add_type(&mut self, type_: Type) -> Result<()>

Registers a new type with the database. This is unstable, and will likely need to be moved to the initialization of the database object in order to allow reconnecting to an existing database.

fn get_type(&self, type_str: &str) -> Option<Type>

Looks for a named type in the database's registry. This function is primarily useful for the DSL shorthand for constructing queries, since it allows you to use names of types when declaring functions rather than type objects.

fn get_predicate(&self, pred_name: &str) -> Option<&Predicate>

Fetches a predicate by name

fn new_predicate(&mut self, pred: &Predicate) -> Result<()>

Persists a predicate by name

fn search_facts(&self, query: &Vec<Clause>) -> Result<Vec<Vec<Value>>>

Attempt to match the right hand side of a datalog rule against the database, returning a list of solution assignments to the bound variables.