Struct holmes::mem_db::MemDB

pub struct MemDB {
    // some fields omitted
}

MemDB is an in-memory mock up of the fact database interface.

While it can be useful for quick tests, it should not be depended on for anything serious, even if you want a standalone app. It is very slow and persists nothing.

Methods

impl MemDB

fn new() -> MemDB

Creates a fresh empty MemDB.