Trait holmes::pg::dyn::HashTO

pub trait HashTO {
    fn hash_to(&self, _: &mut Hasher);
}

HashTO is a Trait Object-safe version of the Hash trait, using a reference to a Hasher rather than a polymorphic one in order to allow construction of trait objects.

HashTO is implemented automatically for any type implementing Hash, allowing Trait Objects to implement Hash. For example:

use holmes::pg::dyn::HashTO;
use std::hash::{Hash, Hasher};
trait Foo : HashTO {}
#[derive(Hash)]
struct Bar;
impl Foo for Bar {}
impl Hash for Foo {
  fn hash<H : Hasher>(&self, hasher : &mut H) {
    self.hash_to(hasher)
  }
}

This trick allows Value and Type trait objects to be used as keys in maps by having types implementing the ValueT or TypeT interface derive Hash.

Required Methods

fn hash_to(&self, _: &mut Hasher)

hash_to captures the same functionality as Hash's hash(), but in a trait object safe way.