Module holmes::pg::dyn::types

This module defines the trait new types must implement, along with several core types to avoid the need to rewrite basic types every time. It is heavily codependent on the values module.

Structs

Bool
Boolean type

Bytes
Bytes is for storing raw data If you want to store text, use the String type.

LargeBytes
LargeBytes is for storing raw data which should not be considered in uniqueness checks If you want to store text, use the String type.

LargeString

List
A list of another Type This type is anonymous.

String
String type Use this for text. If you want to store a buffer, use Bytes instead.

Trap
The Trap type is used to represent types not yet present in the program, but present in the database. Most forms of interaction with a trap type will take down the program.

Tuple
A tuple of other Types This type is anonymous.

UInt64
Unsigned 64-bit int type

Traits

TypeT
The TypeT trait defines the interface a new Holmes type must implement to be registered.

Functions

default_types
Provides a list of provided named types for use by the database as a default set of types.