Trait holmes::pg::dyn::types::TypeT

pub trait TypeT: Sync + Send + HashTO + Any {
    fn name(&self) -> Option<&'static str>;
    fn extract(&self, &mut RowIter) -> Value;
    fn repr(&self) -> Vec<String>;
    fn inner(&self) -> &Any;
    fn inner_eq(&self, other: &TypeT) -> bool;
}

The TypeT trait defines the interface a new Holmes type must implement to be registered.

Required Methods

fn name(&self) -> Option<&'static str>

For a registered type, name() will provide the way to add it to a predicate, and the thing to pattern match against when loading from the db. None will be returned if the type is anonymous, such as a tuple or list.

fn extract(&self, &mut RowIter) -> Value

Takes in an iterator over the row, then attempts to read a value of the type specified.

fn repr(&self) -> Vec<String>

Generates the database representation of the fields required. For example, for a bitvector this would be vec!["bytea".to_string(), "int64".to_string()] or similar, depending on how you chose to construct the size.

fn inner(&self) -> &Any

Returns a dynamic representation of the trait object.

Trait objects cannot be cast to other trait objects, even if they implement the other trait necessarily. In order to access functions under the Any trait, I have the non-trait-object implementation provide access to its &Any representation.

fn inner_eq(&self, other: &TypeT) -> bool

Check equality

Similar to inner, inner_eq exports a PartialEq instance from the underlying type.