Module holmes::pg::dyn::values

This module defines the trait new values must implement, along with several core values to instantiate the basic types provided in types. It is heavily codependent on the values module.

Structs

Bool
Holds a boolean

Bytes
Holds raw data

LargeBWrap
Wrapper newtype pattern for buffers which are too large to be reasonably indexed or matched on.

LargeBytes
Holds large raw data - if your buffer is larger than 256 bytes, you probably want to use this rather than Bytes

List
A list of samely typed values.

String
Holds text

Tuple
A tuple of potentially differently typed values.

UInt64
Holds an unsigned 64-bit int

Statics

FILE_CACHE

Traits

ToValue
Represents the transformability of some type into a dynamic value.

ValueT
This trait defines the interface any value must implement in order to be used in the Holmes language.