Struct holmes::pg::dyn::values::LargeBytes

pub struct LargeBytes { /* fields omitted */ }

Holds large raw data - if your buffer is larger than 256 bytes, you probably want to use this rather than Bytes

Methods

impl LargeBytes

fn new(val: Vec<u8>) -> Arc<Self>

Creates a new Holmes value holding raw data.

fn from_hash(hash: &str) -> Arc<Self>

Generate a LargeBytes value from its hash if already stored. This function does not have any error handling, so it should only be used if the user is certain the value has already been stored.