Struct holmes::Holmes

pub struct Holmes {
    // some fields omitted
}

Encapsulates the user-level interface to a holmes database

Methods

impl Holmes

fn new(db: DB) -> Result<Holmes>

Create a new holmes instance from a db specification

fn destroy(self) -> Result<()>

Tear down and destroy the database. THIS DELETES YOUR DATA

fn add_predicate(&mut self, pred: &Predicate) -> Result<()>

Register a new predicate with the database. This should be done before adding facts to that predicate or rules that refer to it. Do not attempt to register two predicates with the same name.

fn add_fact(&mut self, fact: &Fact) -> Result<()>

Add a new fact to the database. The predicate must be registered first. Duplicate facts are ignored.

fn query(&mut self, query: &Vec<Clause>) -> Result<Vec<Vec<Value>>>

Query the database with the right hand side of a datalog rule, returning all legal assignments to the variables to make it true

fn get_type(&self, name: &str) -> Option<Type>

Get a dynamic type registered with Holmes by name. If it is present, it returns Some(Type), otherwise it returns None. The type database is prepopulated with a few defaults, but otherwise this will only work if you registered the type via add_type.

fn add_type(&mut self, type_: Type) -> Result<()>

Registers a new type with Holmes. It must be a named type, e.g. type.name() would return the Some branch. Types should be registered exactly once

fn add_rule(&mut self, rule: &Rule) -> Result<()>

Adds a new inference rule to the Holmes program

fn reg_func(&mut self, name: String, input_type: Type, output_type: Type, func: Box<Fn(Value) -> Value>) -> Result<()>

Registers a dynamically typed function for use in Holmes where clauses