Trait holmes::pg::dyn::types::TypeT

pub trait TypeT: HashTO + Any {
    fn name(&self) -> Option<&'static str>;

    fn extract(&self, _: &mut RowIter) -> Option<Value>;

    fn repr(&self) -> &'static str;

    fn inner(&self) -> &Any;

    fn inner_eq(&self, other: &TypeT) -> bool;

    fn large(&self) -> bool;
}

The TypeT trait defines the interface a new Holmes type must implement to be registered.

Required Methods

fn name(&self) -> Option<&'static str>

For a registered type, name() will provide the way to add it to a predicate, and the thing to pattern match against when loading from the db. None will be returned if the type is anonymous, such as a tuple or list.

fn extract(&self, _: &mut RowIter) -> Option<Value>

Takes in an iterator over the row, then attempts to read a value of the type specified.

fn repr(&self) -> &'static str

Generates the database representation of the field required.

fn inner(&self) -> &Any

Returns a dynamic representation of the trait object.

Trait objects cannot be cast to other trait objects, even if they implement the other trait necessarily. In order to access functions under the Any trait, I have the non-trait-object implementation provide access to its &Any representation.

fn inner_eq(&self, other: &TypeT) -> bool

Check equality

Similar to inner, inner_eq exports a PartialEq instance from the underlying type.

fn large(&self) -> bool

List of subindexes to be ignored when checking uniqueness. Intended to be used to ignore large payloads in favor of hashes