Trait holmes::pg::dyn::values::ValueT

pub trait ValueT: HashTO + Debug + Display + Any {
    fn type_(&self) -> Type;

    fn get(&self) -> &Any;

    fn to_sql(&self) -> Vec<&ToSql>;

    fn inner(&self) -> &Any;

    fn inner_eq(&self, other: &ValueT) -> bool;

    fn inner_ord(&self, _: &ValueT) -> Option<Ordering>;
}

This trait defines the interface any value must implement in order to be used in the Holmes language.

Required Methods

fn type_(&self) -> Type

Returns the type of the value This is needed if to do type checking, or tuple values

fn get(&self) -> &Any

Get a rust dynamic type version of the value contained.

Since Holmes is a typed language, the user of this function should know what to expect and be able to cast from Any into it.

fn to_sql(&self) -> Vec<&ToSql>

Converts the value into a list of ToSql trait objects to allow for insertion into the database via a prepared query

fn inner(&self) -> &Any

Returns a dynamic representation of the trait object.

Trait objects cannot be cast to other trait objects, even if they implement the other trait necessarily. In order to access functions under the Any trait, I have the non-trait-object implementation provide access to its &Any representation.

fn inner_eq(&self, other: &ValueT) -> bool

Check equality

Similar to inner, inner_eq exports a PartialEq instance from the underlying type.

fn inner_ord(&self, _: &ValueT) -> Option<Ordering>

Check order

Similar to inner, inner_ord exports an Ord instance form the underlying type