Struct holmes::pg::RowIter

pub struct RowIter<'a> { /* fields omitted */ }

An iterator over a postgres::rows::Row. It does not implement the normal iter interface because it does not have a set item type, but it implements a similar interface for ease of use.

Methods

impl<'a> RowIter<'a>

fn new(row: &'a Row) -> Self

Create a new row iterator starting at the beginning of the provided row

fn next<T>(&mut self) -> Option<T> where
T: FromSql,

Gets the next item in the row, using a FromSql instance to read it. If there is not a next item, returns None